Applying our local modifications to server.xml
The <Connector> settings are needed because we have Apache terminate SSL. See http://confluence.atlassian.com/display/JIRA/Integrating+JIRA+with+Apache+using+SSL for more details.

The <Loader> setting is needed for us to override email templates.

The access log is disabled because we have apache take them for consistency with other things we do.